[Requirement #9782] - The Workstation will continue to support basic DICOM functionality as a PACS viewer
Added some changes to support deletion of WorkItems that are canceled or
complete.  Also slightly changed the WorkItemProcessor so its a singleton, so
we can call it to cancel a WorkItem.

+++ ImageViewer/Shreds/WorkItemService/ProcessorThreadPool.cs
@@ -198,7 +198,7 @@ namespace ClearCanvas.ImageViewer.Shreds.WorkItemService
 			                       	});
 		}
 
-        public void Cancel(int workQueueOid)
+        public void Cancel(long workQueueOid)
         {
             foreach (WorkItemThreadParameter queuedItem in _queuedItems)
             {
